requirements.txt
requests
httpx
beautifulsoup4
pyfiglet
rich
pytest
python-whois
dnspython